Fixed duplicate tweet parsing
The code can now store tweets and users and sentiment in a MySQL
database. Each tweet is now only parsed once….mostly. Sometimes the
most recently known tweet and the newest unknown tweet are the same and
are therefore duplicated. Will work on this later. Also, the check to
see if a tweet is more than 3 words long breaks as well. Need to fix
that as well.